Trait rheap::Heap

pub trait Heap<T>where
    T: Ord + Eq + Copy,{
    fn insert(&mut self, element: T);
    fn extract(&mut self) -> Option<T>;
    fn find(&self, element: &T) -> Option<usize>;
    fn update<F>(&mut self, index: usize, update_func: F) -> Result<()>
    where
        F: Fn(&mut T);
    fn remove(&mut self, index: usize) -> Result<T>;
    fn len(&self) -> usize;
    fn clear(&mut self);
    fn is_empty(&self) -> bool;
}

Trait that describes a minimum or maximum heap.

Required Methods

fn insert(&mut self, element: T)

Inserts an element into the heap.

fn extract(&mut self) -> Option<T>

Removes the element from the top of the heap. Returns None if the heap is empty.

fn find(&self, element: &T) -> Option<usize>

Performs a linear search to find the index of element on the heap. If element is found, then it will return its index on the heap. Otherwise, it will return None.

fn update<F>(&mut self, index: usize, update_func: F) -> Result<()>where
F: Fn(&mut T),

Updates the value of the element at index. Returns and error if the element is not found or the index is out of bounds.

fn remove(&mut self, index: usize) -> Result<T>

Removes and returns the element at index. Returns an error if the heap is empty or if the index is out of bounds.

fn len(&self) -> usize

Returns the number of elements in the heap, also referred to as its ‘length’.

fn clear(&mut self)

Clears the heap, removing all elements. Note that this method has no effect on the allocated capacity of the heap.

fn is_empty(&self) -> bool

Returns true if the heap contains no elements.